Implemented basic loading of config from ini file.
The previous configuration was purely Python based: the config file
was literally a module. The variables could be altered at the
command line and were then loaded to the file. Now, the variables
are completely extracted to the config.ini file, and an example is
provided in the defaults section. The program will try to access
this config first by an environment variable specified in
ENVIRONMENT_VARIABLE_NAME. Otherwise, it does a local search in the
current directory at the DEFAULT_CONFIGURATION_FILE. The most
important section of the configuration is the Paths section.

Also, logging with levels is now supported, although not fully
utilized yet.