Added configuration verification. Moved config.
Before, there was no core enforcement that the configuration contained all the
required paths. Now a listing of the missing paths are logged as errors and
then an Error if thrown. The paths required should be listed in the
example/config.ini file, and this file is now required to be located in this
position relative to the config.py file.