Introduced SafeDictionary class
While objectifying the student class, I kept running into the issue
that I wanted the keys to be 'safe' for direct application as
attributes on the student object. However, I also needed access to
the keys before this transformation. Thus, I made a dictionary class
that does this operation for me. Furthermore, there is a decorator
that ensures the dictionaries of a function are all
SafeDictionaries.