Implemented better unittesting and configurations
Now configuration also includes the location of a grading configuration, which
only points to a WebAssign config. The remainder of grading configs will be
loaded locally to a category file, to be implemented later. There are a myriad
of unittest classes available in the tests.setup module, and the appropriate
tests for them. Furthermore, efficient testing and setup required a uniform
initialization script, which is now implemented and tested.