Fixed Exception text in Canvas gradebook generation.
I kept getting an exception that the gradebook couldn't find a WA
id for a particular student. After checking the WA file and the
code, I found that the exception text was actually wrong: the issue
is that the Canvas ID for that student wasn't found (due to the
student withdrawing from the course).